+++ ChangeLog.md
@@ -3,6 +3,7 @@
 ## Apr 29, 2022
 * pyUI/ui.py: Add binders (both positive and negative) for joint sliders to control several joints together.
 * pyUI/ui.py: Can control multiple robots at once. 
+* ardSerial: Get return value from threads.
 
 ## Apr 24, 2022
 * pyUI/ui.py: Support loading and sending large angles